// src/interfaces/local.ts
export interface Local {
    id: string;
    numero_del_local: number;
    nombre: string;
    estado: string;
    estado_sql_login: string;
    marcas: string;
    operativa:boolean;
}
/*
 {
    "local_id": "1fd9e765-a419-4873-aa7b-fac2e57ad160",
    "numero_del_local": 409,
    "local_nombre": "P.S. EU MAIPU",
    "marcas": "Pedro Juan y Diego, Pollo Stop",
    "cliente_nombre": "Unifood",
    "estado": "OK"
  },
*/